build.gradle: Pass the main assets directory path to CMake
In order to copy the assets from the CMake sources folder to the
assets folder, it is essential to know its path.

Yet ! Android does not provide this information automatically.
The CMake variable ANDROID_ASSETS_DIRECTORIES is not set by Android
during the build.

So we provide this information manually, expecting that "someday", the
Android build system will set this common variable automatically.

Note that ANDROID_ASSETS_DIRECTORIES is an official CMake variable.
See :
https://cmake.org/cmake/help/latest/prop_tgt/ANDROID_ASSETS_DIRECTORIES.html

Signed-off-by: Myy <myy@miouyouyou.fr>