Repaired the Android implementation...
Which reminded me how I HATE the Android build system.
Gradle is PURE SHIT when it comes to having a build system that works
over time.
After a few months, gradle start to annoy you because "You're not using
the latest trendy gradle. So 'has-been'. You should reaaallyy
upgrade, you know ? Well, if you want to be a LOSER, just add some
stupid environment variable with the following value !"

Then you upgrade the Gradle plugin, which annoys you because you're not
using the latest Gradle, then you upgrade gradle, and then gradle
whines because there's YET another newer version of the plugin, then
you upgrade the plugin, then you cannot compile anything anymore
because the old tasks names are "ambiguous"...

And THEN you discover how stupid this build system is, when combined
with CMake AND the fact that the Android NDK is NOT C11 compatible.
Meaning that aligned_alloc is NOT available. Fortunately,
posix_memalign is available so... you can reimplement aligned_alloc
with posix_memalign just in case.

Anyway, this took WAY more time than needed. However, that allowed
me to repair a few things here and there, and get a few things better.
So it's not a complete loss.

Still, I REALLY need to find another way to generate the NativeActivity
and package the whole thing without the Android tools.

Signed-off-by: Miouyouyou (Myy) <myy@miouyouyou.fr>