The android input system is working !
Yay...

After looking around, in order to understand why the fuck is it
impossible to get the soft input twice on an EditText, it seems
that I was right. The window focus is essential.

In order to get the window focus, as far as I understand, by
looking at the code, you first you need to avoid the
"touchMode", as that generates various additional checks, and
then you need to provoke a requestFocus(Focus.DOWN, null),
followed by a onWindowFocusChanged(true).
Using both, the InputMethodManager should receive a focusIn
and the keyboard should come back. I'll have to test this.

Signed-off-by: Miouyouyou (Myy) <myy@miouyouyou.fr>